rename verify pkg to kverify
